refactor: move response lock coordination to lifecycle
